Release 1.5.0 (24th Aug 2018)
    Windows Support for Producer SDK
    Intermittent producer support
    Various bug fixes

+++ kinesis-video-pic/src/client/src/AckParser.cpp
@@ -38,7 +38,7 @@ STATUS parseFragmentAck(PKinesisVideoStream pKinesisVideoStream, UPLOAD_HANDLE u
 
     // If ack segment is specified and ack segment size is 0 then we should get the C-string size
     if (0 == ackSegmentSize) {
-        ackSegmentSize = STRNLEN(ackSegment, MAX_ACK_FRAGMENT_LEN);
+        ackSegmentSize = (UINT32) STRNLEN(ackSegment, MAX_ACK_FRAGMENT_LEN);
     } else {
         CHK(ackSegmentSize <= MAX_ACK_FRAGMENT_LEN, STATUS_INVALID_ACK_SEGMENT_LEN);
     }
